requirements.txt
streamlit>=1.33.0
sentence-transformers>=2.6.1
scikit-learn>=1.4.2
nltk>=3.8.1
pymupdf>=1.24.1
python-docx>=1.1.0
pandas>=2.2.2
numpy>=1.26.4
fpdf2>=2.7.9
